refactor: 🏗️ rename all model files to follow .model.js naming convention
