+++ server/models/disease.modal.js
@@ -1,6 +1,6 @@
 const { DataTypes } = require("sequelize");
-const { sequelize } = require("../config/dbConnection");
-const User = require("./userModel");
+const { sequelize } = require("../config/db.config");
+const User = require("./user.modal");
 
 const Disease = sequelize.define(
   "Disease",
