+++ server/models/status.modal.js
@@ -1,5 +1,5 @@
 const { DataTypes } = require("sequelize");
-const { sequelize } = require("../config/dbConnection");
+const { sequelize } = require("../config/db.config");
 
 const Status = sequelize.define(
   "Status",
